+++ src/services/DecoratorService.ts
@@ -20,12 +20,6 @@ export class DecoratorService {
   private static isHighlighted = false;
 
   public static register() {
-    const borderColor =
-      Extension.getInstance().getSetting<string>(Config.highlight.borderColor) || "rgba(255, 0, 0, 0.5)";
-    const background =
-      Extension.getInstance().getSetting<string>(Config.highlight.background) ||
-      "var(--vscode-editor-selectionBackground)";
-
     let blur = Extension.getInstance().getSetting<number>(Config.highlight.blur) || 0;
     if (blur < 0) {
       blur = 0;
@@ -38,46 +32,16 @@ export class DecoratorService {
       opacity = 1;
     }
 
-    const borderStyles = {
-      borderColor,
-      borderStyle: "solid;",
-    };
-
-    const genericStyles: DecorationRenderOptions = {
-      isWholeLine: true,
-      backgroundColor: background,
-      ...borderStyles,
-    };
-
-    DecoratorService.lineDecorator = window.createTextEditorDecorationType({
-      ...genericStyles,
-      borderWidth: "2px;",
-    });
-
-    DecoratorService.startBlockDecorator = window.createTextEditorDecorationType({
-      ...genericStyles,
-      textDecoration: "none;",
-      borderWidth: "2px 2px 0 2px",
-      opacity: "1; filter: blur(0);",
-    });
-
-    DecoratorService.betweenBlockDecorator = window.createTextEditorDecorationType({
-      ...genericStyles,
-      borderWidth: "0 2px 0 2px",
-      opacity: "1; filter: blur(0);",
-    });
-
-    DecoratorService.endBlockDecorator = window.createTextEditorDecorationType({
-      ...genericStyles,
-      textDecoration: "none;",
-      borderWidth: "0 2px 2px 2px",
-      opacity: "1; filter: blur(0);",
-    });
-
     const opacityAndBlur = `${opacity}; filter: blur(${blur}px);`;
     DecoratorService.blurDecorator = window.createTextEditorDecorationType({
       opacity: opacityAndBlur,
     });
+
+    // Initialize decorators
+    DecoratorService.setLineDecorator();
+    DecoratorService.setBeforeDecorators();
+    DecoratorService.setBetweenDecorators();
+    DecoratorService.setAfterDecorators();
 
     // Remove the highlight when the user clicks in the editor
     window.onDidChangeTextEditorSelection((e) => {
@@ -95,11 +59,17 @@ export class DecoratorService {
     DecoratorService.isHighlighted = isDecorated;
   }
 
-  public static hightlightLines(textEditor: TextEditor, range: Range, zoomLevel?: number) {
+  public static hightlightLines(textEditor: TextEditor, range: Range, zoomLevel?: number, isWholeLine?: boolean) {
     const zoomEnabled = Extension.getInstance().getSetting<boolean | number>(Config.highlight.zoom);
 
     // Remove the previous highlight
     DecoratorService.unselect(textEditor);
+
+    // Reset the decorators
+    DecoratorService.setLineDecorator();
+    DecoratorService.setBeforeDecorators();
+    DecoratorService.setBetweenDecorators();
+    DecoratorService.setAfterDecorators();
 
     if (typeof zoomLevel !== "undefined" || zoomEnabled) {
       DecoratorService.isZoomed = true;
@@ -119,26 +89,62 @@ export class DecoratorService {
 
     // Get before and after lines
     const beforeLine = range.start.line;
-    const afterLine = range.end.line + 1;
+    const afterLine = range.end.line;
 
-    // Set the blur on the before and after lines
+    const beforeChar = range.start.character === 0 ? 0 : range.start.character;
+    const afterChar = range.end.character === 0 ? 0 : range.end.character;
+    const line = textEditor.document.lineAt(afterLine);
+
+    // Set the blur on the before lines
     let blurRanges = [];
-    if (beforeLine >= 0) {
-      const beforeRange = new Range(0, 0, beforeLine, 0);
+    if (beforeLine >= 0 || beforeChar > 0) {
+      let endCharacter = isWholeLine ? 0 : beforeChar;
+      if (beforeChar === 0) {
+        endCharacter = 0;
+      }
+
+      const beforeRange = new Range(0, 0, beforeLine, endCharacter);
       blurRanges.push(beforeRange);
     }
 
+    // Set the blur on the after lines
     if (afterLine < textEditor.document.lineCount) {
-      const afterRange = new Range(afterLine, 0, textEditor.document.lineCount, 0);
+      let afterRange = new Range(afterLine + 1, 0, textEditor.document.lineCount, 0);
+      // Check if the last character is not the end of the line
+      // If it is the end, the next line can be blurred
+      // We add 1 to the range to include the end of the line
+      if (afterChar !== line.range.end.character + 1) {
+        // It is not the end of the line, check if the whole line needs to be blurred
+        if (!isWholeLine) {
+          afterRange = new Range(afterLine, afterChar || 0, textEditor.document.lineCount, 0);
+        }
+      }
       blurRanges.push(afterRange);
     }
 
+    // Set the blur decorator
     textEditor.setDecorations(DecoratorService.blurDecorator, blurRanges);
 
     if (range.start.line === range.end.line) {
+      DecoratorService.setLineDecorator(beforeChar > 0 && isWholeLine);
       textEditor.setDecorations(DecoratorService.lineDecorator, [range]);
     } else {
-      const startRange = new Range(range.start.line, 0, range.start.line, 0);
+      let enableWholeLine = true;
+      let startRange = new Range(range.start.line, 0, range.start.line, 0);
+      if (!isWholeLine) {
+        enableWholeLine = !(beforeChar > 0);
+        DecoratorService.setBeforeDecorators(enableWholeLine);
+        DecoratorService.setBetweenDecorators(enableWholeLine);
+      }
+
+      if (beforeChar > 0) {
+        // Get line length of the start line
+        const line = textEditor.document.lineAt(range.start.line);
+        const lineLength = line.text.length;
+
+        startRange = new Range(range.start.line, beforeChar, range.start.line, lineLength);
+      }
+
       textEditor.setDecorations(DecoratorService.startBlockDecorator, [startRange]);
 
       const nextLine = range.start.line + 1;
@@ -147,7 +153,12 @@ export class DecoratorService {
         textEditor.setDecorations(DecoratorService.betweenBlockDecorator, [betweenRange]);
       }
 
-      const endRange = new Range(range.end.line, 0, range.end.line, 0);
+      let endRange = new Range(range.end.line, 0, range.end.line, 0);
+      if (!isWholeLine) {
+        DecoratorService.setAfterDecorators(enableWholeLine);
+        endRange = new Range(range.end.line, 0, range.end.line, afterChar);
+      }
+
       textEditor.setDecorations(DecoratorService.endBlockDecorator, [endRange]);
     }
 
@@ -166,5 +177,89 @@ export class DecoratorService {
       DecoratorService.isZoomed = false;
       commands.executeCommand("editor.action.fontZoomReset");
     }
+  }
+
+  private static getGenericStyles(): DecorationRenderOptions {
+    const borderColor =
+      Extension.getInstance().getSetting<string>(Config.highlight.borderColor) || "rgba(255, 0, 0, 0.5)";
+    const background =
+      Extension.getInstance().getSetting<string>(Config.highlight.background) ||
+      "var(--vscode-editor-selectionBackground)";
+
+    const borderStyles = {
+      borderColor,
+      borderStyle: "solid;",
+    };
+
+    const genericStyles: DecorationRenderOptions = {
+      isWholeLine: true,
+      backgroundColor: background,
+      ...borderStyles,
+    };
+    return genericStyles;
+  }
+
+  private static setLineDecorator(isWholeLine = true) {
+    const genericStyles = DecoratorService.getGenericStyles();
+    const lineStyles = {
+      ...genericStyles,
+      borderWidth: "2px;",
+      isWholeLine,
+    };
+
+    if (isWholeLine) {
+      lineStyles.borderColor = "transparent";
+    }
+
+    DecoratorService.lineDecorator = window.createTextEditorDecorationType(lineStyles);
+  }
+
+  private static setBeforeDecorators(isWholeLine = true) {
+    const genericStyles = DecoratorService.getGenericStyles();
+    const lineStyles = {
+      ...genericStyles,
+      textDecoration: "none;",
+      borderWidth: "2px 2px 0 2px",
+      opacity: "1; filter: blur(0);",
+      isWholeLine,
+    };
+
+    if (!isWholeLine) {
+      lineStyles.borderColor = "transparent";
+    }
+
+    DecoratorService.startBlockDecorator = window.createTextEditorDecorationType(lineStyles);
+  }
+
+  private static setBetweenDecorators(isWholeLine = true) {
+    const genericStyles = DecoratorService.getGenericStyles();
+    const lineStyles = {
+      ...genericStyles,
+      borderWidth: "0 2px 0 2px",
+      opacity: "1; filter: blur(0);",
+    };
+
+    if (!isWholeLine) {
+      lineStyles.borderColor = "transparent";
+    }
+
+    DecoratorService.betweenBlockDecorator = window.createTextEditorDecorationType(lineStyles);
+  }
+
+  private static setAfterDecorators(isWholeLine = true) {
+    const genericStyles = DecoratorService.getGenericStyles();
+    const lineStyles = {
+      ...genericStyles,
+      textDecoration: "none;",
+      borderWidth: "0 2px 2px 2px",
+      opacity: "1; filter: blur(0);",
+      isWholeLine,
+    };
+
+    if (!isWholeLine) {
+      lineStyles.borderColor = "transparent";
+    }
+
+    DecoratorService.endBlockDecorator = window.createTextEditorDecorationType(lineStyles);
   }
 }