+++ src/utils/getPositionAndRange.ts
@@ -4,7 +4,7 @@ import { findPositionByLineNumbers, findPositionByPlaceholders } from ".";
 
 /**
  * Retrieves the current position and range based on the provided step.
- * 
+ *
  * @param editor The text document editor.
  * @param step The step object containing the position information.
  * @returns An object with the current position and range.
@@ -12,12 +12,18 @@ import { findPositionByLineNumbers, findPositionByPlaceholders } from ".";
 export const getPositionAndRange = async (
   editor: TextDocument,
   step: Step
-): Promise<{ crntPosition: Position | undefined; crntRange: Range | undefined }> => {
-  let positioning: { position: Position | undefined; range: Range | undefined; } | undefined = undefined;
+): Promise<{
+  crntPosition: Position | undefined;
+  crntRange: Range | undefined;
+  usesPlaceholders: boolean;
+}> => {
+  let positioning: { position: Position | undefined; range: Range | undefined } | undefined = undefined;
+  let usesPlaceholders = false;
 
   if (step.position) {
     positioning = findPositionByLineNumbers(editor, step.position);
   } else if (step.startPlaceholder && step.path) {
+    usesPlaceholders = true;
     positioning = await findPositionByPlaceholders(step.startPlaceholder, step.path, step.endPlaceholder);
   }
 
@@ -25,11 +31,13 @@ export const getPositionAndRange = async (
     return {
       crntPosition: undefined,
       crntRange: undefined,
+      usesPlaceholders,
     };
   }
 
   return {
     crntPosition: positioning.position,
-    crntRange: positioning.range
+    crntRange: positioning.range,
+    usesPlaceholders,
   };
 };